analyzer: split boolean.h properly.
