+++ analyzer/types/boolean.cpp
@@ -21,7 +21,9 @@
  **/
 
 #include "vapor/analyzer/types/boolean.h"
+#include "vapor/analyzer/expressions/expression.h"
 #include "vapor/analyzer/symbol.h"
+#include "vapor/analyzer/variables/boolean.h"
 #include "vapor/codegen/ir/type.h"
 #include "vapor/codegen/ir/variable.h"
 
@@ -32,20 +34,6 @@ inline namespace _v1
     void boolean_type::_codegen_type(ir_generation_context &) const
     {
         _codegen_t = codegen::ir::builtin_types().boolean;
-    }
-
-    variable_ir boolean_constant::_codegen_ir(ir_generation_context &) const
-    {
-        return { codegen::ir::value{ codegen::ir::boolean_value{ _value } } };
-    }
-
-    statement_ir boolean_literal::_codegen_ir(ir_generation_context &) const
-    {
-        return { codegen::ir::instruction{ none,
-            none,
-            { boost::typeindex::type_id<codegen::ir::pass_value_instruction>() },
-            {},
-            codegen::ir::value{ codegen::ir::boolean_value{ _value->get_value() } } } };
     }
 
     std::unique_ptr<type> make_boolean_type()
